+++ resources/js/components/shared/LaunchpadSettingsModal.vue
@@ -1,5 +1,6 @@
 <template>
   <modal
+    id="launchpadSettingsModal"
     ref="my-modal-save"
     size="lg"
     class="modal-dialog modal-dialog-centered"
@@ -17,248 +18,8 @@
             md="12"
             class="no-padding"
           >
-            <div class="d-flex align-items-center w-100">
-              <label>{{ $t("Launchpad Carousel") }}</label>
-              <input
-                ref="fileInput"
-                type="file"
-                style="display: none"
-                accept="image/*"
-                @change="handleImageUpload"
-              >
-            </div>
-          </div>
-          <div
-            ref="thumbnailsContainer"
-            class="image-thumbnails-container"
-            @drop="handleDrop"
-            @dragover.prevent
-            @dragstart.prevent="handleDragStart"
-          >
-            <div
-              v-if="images.length === 0 && !loadingImage"
-              md="12"
-              class="text-center images-info images-container"
-            >
-              <div
-                class="drag-and-drop-container"
-              >
-                <i class="fas fa-image" />
-                <div>
-                  {{ $t("Formats: PNG, JPG. 2 MB") }}
-                </div>
-                {{ $t("Recommended: 1800 x 750 px") }}
-              </div>
-            </div>
-            <div
-              v-else
-              class="images-container"
-            >
-              <div
-                v-for="(image, index) in images"
-                :key="index"
-                md="6"
-              >
-                <div
-                  v-if="image.type !== 'embed'"
-                  class="d-flex justify-content-end align-items-end thumbnail image-style mr-2"
-                  @mouseover="showDeleteIcon(index)"
-                  @mouseleave="hideDeleteIcon(index)"
-                >
-                  <div
-                    v-if="showDeleteIcons[index] || focusIcons[index]"
-                    :id="`popover-button-event-${index}`"
-                    class="delete-icon"
-                    @click="focusIcon(index)"
-                  >
-                    <div>
-                      <i class="fas fa-trash-alt p-0 custom-color" />
-                    </div>
-                    <b-popover
-                      ref="popover"
-                      :show.sync="focusIcons[index]"
-                      :target="`popover-button-event-${index}`"
-                      triggers="focus"
-                      placement="bottom"
-                    >
-                      <div class="d-flex popover-custom">
-                        <p class="m-0">
-                          {{ $t("Do you want to delete this image?") }}
-                        </p>
-                        <button
-                          type="button"
-                          class="btn btn-delete-image btns-popover"
-                          @click="deleteImage(index)"
-                        >
-                          {{ $t("Delete") }}
-                        </button>
-                        <button
-                          type="button"
-                          class="btn btn-cancel-delete btns-popover"
-                          @click="unfocusIcon(index)"
-                        >
-                          {{ $t("Cancel") }}
-                        </button>
-                      </div>
-                    </b-popover>
-                  </div>
-                  <img
-                    v-if="image.url"
-                    :src="image.url"
-                    :alt="$t('No Image')"
-                    class="image-style"
-                  >
-                </div>
-                <div
-                  v-else
-                  id="embedFile"
-                >
-                  <div
-                    :id="`popover-embed-event-${index}`"
-                    class="square-image image-style"
-                    @click="focusIcon(index)"
-                  >
-                    <i class="fas fa-link" />
-                  </div>
-                  <b-popover
-                    ref="popover"
-                    :show.sync="focusIcons[index]"
-                    :target="`popover-embed-event-${index}`"
-                    triggers="focus"
-                    placement="bottom"
-                  >
-                    <div class="popover-embed">
-                      <label>
-                        {{ $t("Embed URL") }}
-                      </label>
-                      <input
-                        :id="`embed-input-${index}`"
-                        v-model="embedUrls[index]"
-                        class="form-control input-custom"
-                        type="text"
-                        rows="5"
-                        :aria-label="$t('Embed URL')"
-                      />
-                      <div
-                        v-if="!deleteEmbed"
-                        class="d-flex justify-content-between"
-                      >
-                        <i
-                          class="fas fa-trash-alt custom-trash-icon"
-                          @click="deleteEmbed = true"
-                        />
-                        <div class="d-flex">
-                          <button
-                            type="button"
-                            class="btn btn-cancel-delete btns-popover"
-                            @click="cancelEmbed(index)"
-                          >
-                            {{ $t("Cancel") }}
-                          </button>
-                          <button
-                            type="button"
-                            class="btn btn-delete-image btns-popover"
-                            @click="saveEmbed(index)"
-                          >
-                            {{ $t("Apply") }}
-                          </button>
-                        </div>
-                      </div>
-                      <div
-                        v-else
-                        class="d-flex justify-content-between"
-                      >
-                        <span
-                          class="text-delete-embed"
-                        >
-                          {{ $t("Delete this embed media?") }}
-                        </span>
-                        <div class="d-flex">
-                          <button
-                            type="button"
-                            class="btn btn-cancel-delete btns-popover"
-                            @click="deleteEmbed = false"
-                          >
-                            {{ $t("Cancel") }}
-                          </button>
-                          <button
-                            type="button"
-                            class="btn btn-delete-embed btns-popover"
-                            @click="deleteEmbedMedia(index)"
-                          >
-                            {{ $t("Delete") }}
-                          </button>
-                        </div>
-                      </div>
-                    </div>
-                  </b-popover>
-                </div>
-              </div>
-              <div
-                v-if="loadingImage"
-                class="square-image image-style"
-              >
-                <i class="fas fa-solid fa-circle-notch fa-spin" />
-              </div>
-            </div>
-            <div
-              v-if="!loadingImage && !notValidImage"
-              id="idDropdownMenuUpload"
-              class="d-flex justify-content-center align-items-center"
-            >
-              <div
-                class="input-file-custom dropdown-toggle"
-                @dragover.prevent
-                type="button"
-                id="dropdownMenuUpload"
-                data-toggle="dropdown"
-                aria-haspopup="true"
-                aria-expanded="false"
-              >
-                <i class="fa fa-plus mr-1" />
-                <span class="font-weight-bold mr-1">
-                  {{ $t("Drag or click here") }}
-                </span>
-                {{ $t("to upload an image") }}
-              </div>
-              <div class="dropdown-menu dropdown-menu-right" aria-labelledby="dropdownMenuUpload">
-                <a
-                  class="dropdown-item"
-                  href="#"
-                  @click="openFileInput"
-                >
-                  {{ $t("Load an Image") }}
-                </a>
-                <a
-                  class="dropdown-item"
-                  href="#"
-                  @click="addEmbedMedia"
-                >
-                  {{ $t("Embed Media") }}
-                </a>
-              </div>
-            </div>
-            <div
-              v-if="loadingImage"
-              class="d-flex justify-content-center align-items-center"
-            >
-              <div
-                class="input-file-custom"
-              >
-                {{ $t("Loading...") }}
-              </div>
-            </div>
-            <div
-              v-if="notValidImage"
-              class="d-flex justify-content-center align-items-center"
-            >
-              <div
-                class="input-file-custom"
-              >
-                <b-icon icon="exclamation-triangle-fill" class="mr-1" variant="warning"></b-icon>
-                {{ $t("Image not valid, try another") }}
-              </div>
-            </div>
+            <label>{{ $t("Launchpad Carousel") }}</label>
+            <input-image-carousel ref="image-carousel" />
           </div>
         </div>
         <div class="options-launchpad">
@@ -320,9 +81,10 @@
 <script>
 import Modal from "./Modal.vue";
 import IconDropdown from "./IconDropdown.vue";
+import InputImageCarousel from "./InputImageCarousel.vue";
 
 export default {
-  components: { Modal, IconDropdown },
+  components: { Modal, IconDropdown, InputImageCarousel },
   props: {
     options: {
       type: Object,
@@ -350,18 +112,13 @@ export default {
   },
   data() {
     return {
-      images: [],
       imagesMedia: [],
-      embedUrls: Array(4).fill(""),
-      showDeleteIcons: Array(4).fill(false),
-      focusIcons: Array(4).fill(false),
       list: {},
       subject: "",
       description: "",
       errors: "",
       selectedSavedChart: "",
       dropdownSavedCharts: [],
-      maxImages: 4,
       processDescription: "",
       processDescriptionInitial: "",
       selectedLaunchpadIcon: "",
@@ -372,19 +129,14 @@ export default {
       processId: "",
       mediaImageId: [],
       dataProcess: {},
-      loadingImage: false,
-      notValidImage: false,
-      deleteEmbed: false,
     };
-  },
-  computed: {
-    
   },
   mounted() {
     this.retrieveSavedSearchCharts();
     this.getDescriptionInitial();
     this.getProcessDescription();
-
+          
+    this.$refs["image-carousel"].setProcessId(this.processId);
     // Receives selected Option from launchpad Icons multiselect
     this.$root.$on("launchpadIcon", this.launchpadIconSelected);
   },
@@ -393,7 +145,6 @@ export default {
      * Get all information related to Launchpad Settings Modal
      */
     getLaunchpadSettings() {
-      this.images = [];
       ProcessMaker.apiClient
         .get(`processes/${this.processId}/media`)
         .then((response) => {
@@ -416,33 +167,14 @@ export default {
           // Load Images into Carousel Container
           const mediaArray = firstResponse.media;
           mediaArray.forEach((media) => {
-            this.convertImageUrlToBase64(media);
+            this.$refs["image-carousel"].convertImageUrlToBase64(media);
           });
-        });
-    },
-    /**
-     * Converts Image from URL to Base64
-     */
-    convertImageUrlToBase64(media) {
-      fetch(media.original_url)
-        .then((response) => response.blob())
-        .then((blob) => {
-          const reader = new FileReader();
-          reader.onloadend = () => {
-            const base64Data = reader.result;
-            this.images.push({ url: base64Data, uuid: media.uuid });
-          };
-          reader.readAsDataURL(blob);
-        })
-        .catch((error) => {
-          console.error("Error loading image:", error);
         });
     },
     showModal() {
       this.subject = "";
       this.description = "";
       this.errors = "";
-      this.images = [];
       this.getLaunchpadSettings();
       this.$refs["my-modal-save"].show();
     },
@@ -459,73 +191,17 @@ export default {
     },
     cleanTabLaunchpad() {
       this.getProcessDescription();
-      this.images = [];
-      this.embedUrls = Array(4).fill("");
       this.retrieveSavedSearchCharts();
+      this.$refs["image-carousel"].cleanCarousel();
       this.showVersionInfo = true;
       this.isSecondaryColor = false;
-      this.notValidImage = false;
-      this.loadingImage = false;
-    },
-    /**
-     * Method to show trash image
-     */
-    showDeleteIcon(index) {
-      return this.$set(this.showDeleteIcons, index, true);
-    },
-    /**
-     * Method to hide trash image
-     */
-    hideDeleteIcon(index) {
-      return this.$set(this.showDeleteIcons, index, false);
-    },
-    /**
-     * Method to focus trash image
-     */
-    focusIcon(index) {
-      this.focusIcons = Array(4).fill(false);
-      this.deleteEmbed = false;
-      this.$set(this.focusIcons, index, true);
-    },
-    /**
-     * Method to unfocus trash image
-     */
-    unfocusIcon(index) {
-      this.$set(this.focusIcons, index, false);
-      this.deleteEmbed = false;
-    },
-    /**
-     * Method to delete image from carousel container
-     */
-    deleteImage(index) {
-      const { uuid } = this.images[index];
-      this.images.splice(index, 1);
-      this.$set(this.showDeleteIcons, index, false);
-      this.$set(this.focusIcons, index, false);
-
-      // Call API to delete
-      ProcessMaker.apiClient
-        .delete(`processes/${this.processId}/media`, {
-          data: { uuid },
-        })
-        .then((response) => {
-          ProcessMaker.alert(this.$t("The image was deleted"), "success");
-        })
-        .catch((error) => {
-          console.error("Error", error);
-        });
-      const params = {
-        indexImage: index,
-        type: "delete",
-      };
-      ProcessMaker.EventBus.$emit("getLaunchpadImagesEvent", params);
     },
     /**
      * Save description field in Process
      */
     saveProcessDescription() {
       if (!this.processDescription) return;
-      this.dataProcess.imagesCarousel = this.images;
+      this.dataProcess.imagesCarousel = this.$refs["image-carousel"].getImages();
       this.dataProcess.launchpad_properties = JSON.stringify({
         saved_chart_id: this.selectedSavedChartId,
         saved_chart_title: this.selectedSavedChart,
@@ -593,13 +269,6 @@ export default {
         });
     },
     /**
-     * Adds index info to dragged object
-     */
-    handleDragStart(event, index) {
-      event.dataTransfer.setData("text/plain", index);
-      event.preventDefault();
-    },
-    /**
      * Method to store initial data from process description field
      */
     getDescriptionInitial() {
@@ -630,141 +299,6 @@ export default {
             this.processDescription = this.processDescriptionInitial;
           }
       }
-    },
-    /**
-     * Method to add image files to thumbnails container
-     */
-    handleImageUpload(event) {
-      if (this.images.length >= this.maxImages) {
-        // The amount of images allowed was reached.
-        ProcessMaker.alert(
-          this.$t("It is not possible to include more than four images."),
-          "danger",
-        );
-        this.$refs.fileInput.value = "";
-        return;
-      }
-      const { files } = event.target;
-      this.handleImages(files);
-      event.target.value = "";
-    },
-    /**
-     * Method that allows drag elements to the container
-     */
-    handleDragOver(event) {
-      event.preventDefault();
-    },
-    /**
-     * This method handles dragged image files and adds each image to list
-     */
-    handleDrop(event) {
-      event.preventDefault();
-
-      // Checks if event has 'dataTransfer' property
-      if (event.dataTransfer) {
-        const { files } = event.dataTransfer;
-
-        // Checks if 'dataTransfer' has 'files' property
-        if (files && files.length > 0) {
-          if (this.images.length + files.length > this.maxImages) {
-            ProcessMaker.alert(
-              this.$t("It is not possible to include more than four images."),
-              "danger",
-            );
-            return;
-          }
-          this.validateImageExtension(files);
-        }
-      }
-    },
-    /**
-     * Generic Method to manage drag and drop and selected images
-     */
-    handleImages(files) {
-      this.validateImageExtension(files);
-    },
-    /**
-     * Adds an image from drag and drop to image container
-     */
-    handleDroppedImage(event) {
-      const { files } = event.dataTransfer;
-      this.handleImages(files);
-    },
-    /**
-     *  Validates images with png and jpg extensions.
-     */
-    validateImageExtension(files) {
-      this.loadingImage = true;
-      Array.from(files).forEach((file) => {
-        if (this.images.length < this.maxImages) {
-          if (this.isValidFileExtension(file.name)) {
-            const reader = new FileReader();
-            reader.onload = (event) => {
-              this.images.push({
-                file,
-                url: event.target.result,
-                uuid: "",
-                type: "image",
-              });
-              this.showDeleteIcons.push(false);
-            };
-            reader.readAsDataURL(file);
-          } else {
-            this.notValidImage = true;
-            setTimeout(() => {
-              this.notValidImage = false;
-            }, 4000);
-            ProcessMaker.alert(
-              this.$t("Only PNG and JPG extensions are allowed."),
-              "danger",
-            );
-          }
-          this.loadingImage = false;
-        }
-      });
-    },
-    addEmbedMedia() {
-      if (this.images.length >= this.maxImages) {
-        // The amount of images allowed was reached.
-        ProcessMaker.alert(
-          this.$t("It is not possible to include more than four images."),
-          "danger",
-        );
-        return;
-      }
-      this.images.push({
-        url: "",
-        uuid: "",
-        type: "embed",
-      });
-      this.focusIcon(this.images.length - 1);
-    },
-    cancelEmbed(index) {
-      this.$set(this.embedUrls, index, "");
-      this.unfocusIcon(index);
-    },
-    saveEmbed(index) {
-      this.images[index].url = this.embedUrls[index];
-      this.unfocusIcon(index);
-    },
-    deleteEmbedMedia(index) {
-      this.embedUrls[index] = '';
-      this.deleteImage(index);
-    },
-    /**
-     * Validate image extensions
-     */
-    isValidFileExtension(fileName) {
-      const allowedExtensions = [".jpg", ".jpeg", ".png"];
-      return allowedExtensions.includes(
-        fileName.slice(fileName.lastIndexOf(".")).toLowerCase(),
-      );
-    },
-    /**
-     * Method to open a screen for image selection from hard drive
-     */
-    openFileInput() {
-      this.$refs.fileInput.click();
     },
     launchpadIconSelected(iconData) {
       this.selectedLaunchpadIcon = iconData.value;
@@ -1004,14 +538,14 @@ b-row, b-col {
   color: #6a7888;
   font-size: 24px;
 }
-#idDropdownMenuUpload .dropdown-toggle::after {
+#launchpadSettingsModal .dropdown-toggle::after {
     display:none;
 }
-#idDropdownMenuUpload .dropdown-menu.show {
+#launchpadSettingsModal .dropdown-menu.show {
   width: 229px;
   padding: 0px;
 }
-#idDropdownMenuUpload .dropdown-item {
+#launchpadSettingsModal .dropdown-item {
   color: #556271;
   padding: 12px;
   font-family: 'Open Sans', sans-serif;
